scraper/requirements.txt
curl_cffi>=0.7.4
supabase>=2.0.0
python-dotenv>=1.0.1
